pipeline: move traits into `traits.rs`
